requirements.txt
openai
requests
python-dotenv
pydantic
apify-client
rich
beautifulsoup4
markdown
lxml
pandas
streamlit
streamlit-chat
streamlit-extras
